[x86_64] F: Add C/C++ Integration Test Support
Add x86_64 C/C++ cross-compilation support and fix IPC wire format for
cross-architecture compatibility between x86 guests and x86_64 host.

FFI and type system:
- Split ffi.rs c_long/c_ulong into conditional 32/64-bit definitions
  matching the LP64 data model on x86_64
- Change stat, dirent, posix_dent, pthread types from repr(C, packed)
  to repr(C) for correct alignment on both architectures

IPC wire format:
- Add fixed-width wire format (always i64) for c_long-based fields in
  timespec and stat serialization, enabling x86 guest to x86_64 host
  communication
- Convert futimens message from transmute-based to explicit byte
  serialization using timespec to_bytes/try_from_bytes
- Update utimensat and fstatat wire format constants
- Add allow(clippy::unnecessary_cast/useless_conversion) for casts
  that are needed on x86 but identity on x86_64

C test portability:
- Fix type-size assertions, struct padding, pointer casts, and aligned
  allocation for LP64 compatibility across all C test suites
- Add x86_64 inline assembly variant for dlfcn-c shared library
- Remove TARGET=x86_64 exclusion guard from generic-guest-binaries.mk

Test configurations:
- Add C/C++ tests to x86_64 single-process and multi-process configs
- Exclude dlfcn-c on x86_64 (requires R_X86_64_* ELF relocations)

Documentation:
- Update x86_64-port.md with toolchain build, FFI types, IPC wire
  format, struct layout, C test portability, and known limitations

+++ src/libs/syscall/src/time/syscall/nanosleep.rs
@@ -13,6 +13,7 @@ use ::sys::{
     },
     time::SystemTime,
 };
+use ::sysapi::ffi::c_long;
 use sysapi::time::timespec;
 
 //==================================================================================================
@@ -98,7 +99,7 @@ pub fn nanosleep(req: &timespec, rem: &mut Option<&mut timespec>) -> Result<(), 
             rem.tv_nsec = 0;
         } else {
             rem.tv_sec = (later.seconds() - now.seconds()) as i64;
-            rem.tv_nsec = (later.nanoseconds() - now.nanoseconds()) as i32;
+            rem.tv_nsec = (later.nanoseconds() - now.nanoseconds()) as c_long;
         }
     }
